[Feature][Transform] optimized code
