+++ seatunnel-e2e/seatunnel-transforms-v2-e2e/seatunnel-transforms-v2-e2e-part-1/src/test/resources/embedding_transform.conf
@@ -62,7 +62,7 @@ source {
 transform {
   Embedding {
     source_table_name = "fake"
-    embedding_model_provider = QIANFAN
+    model_provider = QIANFAN
     model = bge_large_en
     api_key = xxxxxxxx
     secret_key = xxxxxxxx
@@ -78,7 +78,7 @@ transform {
 
   Embedding {
     source_table_name = "fake"
-    embedding_model_provider = DOUBAO
+    model_provider = DOUBAO
     model = ep-20240830113341-wwwqd
     api_key = xxxxxxxx
     api_path = "http://mockserver:1080/v1/doubao/embedding"
@@ -93,7 +93,7 @@ transform {
 
  Embedding {
     source_table_name = "fake"
-    embedding_model_provider = OPENAI
+    model_provider = OPENAI
     model = text-embedding-3-small
     api_key = xxxxxxxx
     api_path = "http://mockserver:1080/v1/openai/embedding"
