+++ seatunnel-transforms-v2/src/main/java/org/apache/seatunnel/transform/nlpmodel/remote/embadding/processor/doubao/DoubaoModel.java
@@ -15,14 +15,14 @@
  * limitations under the License.
  */
 
-package org.apache.seatunnel.transform.embadding.processor.doubao;
+package org.apache.seatunnel.transform.nlpmodel.remote.embadding.processor.doubao;
 
 import org.apache.seatunnel.shade.com.fasterxml.jackson.core.type.TypeReference;
 import org.apache.seatunnel.shade.com.fasterxml.jackson.databind.JsonNode;
 import org.apache.seatunnel.shade.com.fasterxml.jackson.databind.node.ArrayNode;
 import org.apache.seatunnel.shade.com.fasterxml.jackson.databind.node.ObjectNode;
 
-import org.apache.seatunnel.transform.embadding.processor.AbstractModel;
+import org.apache.seatunnel.transform.nlpmodel.remote.embadding.processor.AbstractModel;
 
 import org.apache.http.client.config.RequestConfig;
 import org.apache.http.client.methods.CloseableHttpResponse;
