+++ seatunnel-transforms-v2/src/main/java/org/apache/seatunnel/transform/nlpmodel/remote/llm/LLMTransform.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.seatunnel.transform.llm;
+package org.apache.seatunnel.transform.nlpmodel.remote.llm;
 
 import org.apache.seatunnel.api.configuration.ReadonlyConfig;
 import org.apache.seatunnel.api.table.catalog.CatalogTable;
@@ -26,8 +26,10 @@ import org.apache.seatunnel.api.table.type.SeaTunnelDataType;
 import org.apache.seatunnel.api.table.type.SeaTunnelRow;
 import org.apache.seatunnel.transform.common.SeaTunnelRowAccessor;
 import org.apache.seatunnel.transform.common.SingleFieldOutputTransform;
-import org.apache.seatunnel.transform.llm.model.Model;
-import org.apache.seatunnel.transform.llm.model.openai.OpenAIModel;
+import org.apache.seatunnel.transform.nlpmodel.remote.ModelProvider;
+import org.apache.seatunnel.transform.nlpmodel.remote.ModelTransformConfig;
+import org.apache.seatunnel.transform.nlpmodel.remote.llm.processor.Model;
+import org.apache.seatunnel.transform.nlpmodel.remote.llm.processor.openai.OpenAIModel;
 
 import lombok.NonNull;
 import lombok.SneakyThrows;
@@ -61,18 +63,22 @@ public class LLMTransform extends SingleFieldOutputTransform {
 
     @Override
     public void open() {
-        ModelProvider provider = config.get(LLMTransformConfig.MODEL_PROVIDER);
-        if (provider.equals(ModelProvider.OPENAI)) {
-            model =
-                    new OpenAIModel(
-                            inputCatalogTable.getSeaTunnelRowType(),
-                            outputDataType.getSqlType(),
-                            config.get(LLMTransformConfig.PROMPT),
-                            config.get(LLMTransformConfig.MODEL),
-                            config.get(LLMTransformConfig.API_KEY),
-                            config.get(LLMTransformConfig.OPENAI_API_PATH));
-        } else {
-            throw new IllegalArgumentException("Unsupported model provider: " + provider);
+        ModelProvider provider = config.get(ModelTransformConfig.MODEL_PROVIDER);
+        switch (provider) {
+            case OPENAI:
+                model =
+                        new OpenAIModel(
+                                inputCatalogTable.getSeaTunnelRowType(),
+                                outputDataType.getSqlType(),
+                                config.get(LLMTransformConfig.PROMPT),
+                                config.get(LLMTransformConfig.MODEL),
+                                config.get(LLMTransformConfig.API_KEY),
+                                provider.usedLLMPath(config.get(LLMTransformConfig.API_PATH)));
+                break;
+            case QIANFAN:
+            case DOUBAO:
+            default:
+                throw new IllegalArgumentException("Unsupported model provider: " + provider);
         }
     }
