+++ seatunnel-transforms-v2/src/main/java/org/apache/seatunnel/transform/nlpmodel/remote/llm/LLMTransformConfig.java
@@ -15,21 +15,14 @@
  * limitations under the License.
  */
 
-package org.apache.seatunnel.transform.llm;
+package org.apache.seatunnel.transform.nlpmodel.remote.llm;
 
 import org.apache.seatunnel.api.configuration.Option;
 import org.apache.seatunnel.api.configuration.Options;
 import org.apache.seatunnel.api.table.type.SqlType;
+import org.apache.seatunnel.transform.nlpmodel.remote.ModelTransformConfig;
 
-import java.io.Serializable;
-
-public class LLMTransformConfig implements Serializable {
-
-    public static final Option<ModelProvider> MODEL_PROVIDER =
-            Options.key("model_provider")
-                    .enumType(ModelProvider.class)
-                    .noDefaultValue()
-                    .withDescription("The model provider of LLM");
+public class LLMTransformConfig extends ModelTransformConfig {
 
     public static final Option<SqlType> OUTPUT_DATA_TYPE =
             Options.key("output_data_type")
@@ -43,19 +36,6 @@ public class LLMTransformConfig implements Serializable {
                     .noDefaultValue()
                     .withDescription("The prompt of LLM");
 
-    public static final Option<String> MODEL =
-            Options.key("model")
-                    .stringType()
-                    .noDefaultValue()
-                    .withDescription(
-                            "The model of LLM, eg: if the model provider is OpenAI, the model should be gpt-3.5-turbo/gpt-4o-mini, etc.");
-
-    public static final Option<String> API_KEY =
-            Options.key("api_key")
-                    .stringType()
-                    .noDefaultValue()
-                    .withDescription("The API key of LLM");
-
     public static final Option<Integer> INFERENCE_BATCH_SIZE =
             Options.key("inference_batch_size")
                     .intType()
@@ -67,5 +47,5 @@ public class LLMTransformConfig implements Serializable {
             Options.key("openai.api_path")
                     .stringType()
                     .defaultValue("https://api.openai.com/v1/chat/completions")
-                    .withDescription("The API path of OpenAI");
+                    .withDescription("The API path of OpenAI LLM");
 }
