+++ seatunnel-transforms-v2/src/main/java/org/apache/seatunnel/transform/nlpmodel/remote/llm/processor/Model.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.seatunnel.transform.llm.model;
+package org.apache.seatunnel.transform.nlpmodel.remote.llm.processor;
 
 import org.apache.seatunnel.api.table.type.SeaTunnelRow;
