+++ seatunnel-transforms-v2/src/test/java/org/apache/seatunnel/transform/LLMTransformFactoryTest.java
@@ -17,7 +17,7 @@
 
 package org.apache.seatunnel.transform;
 
-import org.apache.seatunnel.transform.llm.LLMTransformFactory;
+import org.apache.seatunnel.transform.nlpmodel.remote.llm.LLMTransformFactory;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
